8332123: [nmt] Move mallocLimit code to the nmt subdir
