+++ src/hotspot/share/nmt/mallocLimit.cpp
@@ -28,7 +28,7 @@
 #include "nmt/nmtCommon.hpp"
 #include "runtime/java.hpp"
 #include "runtime/globals.hpp"
-#include "services/mallocLimit.hpp"
+#include "nmt/mallocLimit.hpp"
 #include "utilities/globalDefinitions.hpp"
 #include "utilities/parseInteger.hpp"
 #include "utilities/ostream.hpp"
