+++ test/hotspot/gtest/nmt/test_nmt_malloclimit.cpp
@@ -27,7 +27,7 @@
 #include "nmt/memTracker.hpp"
 #include "nmt/nmtCommon.hpp"
 #include "runtime/os.hpp"
-#include "services/mallocLimit.hpp"
+#include "nmt/mallocLimit.hpp"
 #include "utilities/debug.hpp"
 #include "utilities/globalDefinitions.hpp"
 #include "utilities/ostream.hpp"
